refactor(Ray): turned double to float

+++ src/Utils.cpp
@@ -2,14 +2,14 @@
 #include <numbers>
 #include <cmath>
 
-void normalizeAngle(double& angle) {
+void normalizeAngle(float& angle) {
     angle = remainder(angle, 2 * std::numbers::pi);
     if (angle < 0) {
         angle = 2 * std::numbers::pi + angle;
     }
 }
 
-double distanceBetweenPoints(double x1, double y1, double x2, double y2) {
+float distanceBetweenPoints(float x1, float y1, float x2, float y2) {
     return sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
 }
